Moved files to UI folder
