+++ 08 implementation/Mobile App/app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.AppCompat.Light.NoActionBar">
-        <activity android:name=".ThirdActivity"
+        <activity android:name=".UI.ThirdActivity"
             ></activity>
         <activity android:name=".UI.BuildWallActivity">
             <intent-filter>
