+++ 08 implementation/Mobile App/app/src/main/res/layout/activity_basket.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ThirdActivity">
+    tools:context=".UI.ThirdActivity">
 
     <TextView
         android:id="@+id/TextView"
